+++ lib/components/PrettyBytes.vue
@@ -1,0 +1,15 @@
+<script>
+export default {
+
+    props: {
+        value: null,
+    },
+}
+</script>
+
+<template>
+  <span :title="`${value} bytes`">
+    {{ $qordial.formatBytes(value) }}
+  </span>
+
+</template>