+++ vite.config.js
@@ -28,6 +28,7 @@ export default defineConfig({
         // Provide global variables to use in the UMD build
         // for externalized deps
         globals: {
+          pinia: 'pinia',
           vue: 'Vue',
         },
       },